refactored adminpage dir structure
