+++ components/AdminPage/SetPayoutAdminPage/SetPayoutAdminPage.test.js
@@ -2,12 +2,12 @@
  * @jest-environment jsdom
  */
 import React from 'react';
-import { render, screen } from '../../test-utils';
-import SetPayoutAdminPage from '../../components/AdminPage/SetPayoutAdminPage';
-import InitialState from '../../store/Store/InitialState';
-import Constants from '../../test-utils/constant';
+import { render, screen } from '../../../test-utils';
+import SetPayoutAdminPage from '.';
+import InitialState from '../../../store/Store/InitialState';
+import Constants from '../../../test-utils/constant';
 import userEvent from '@testing-library/user-event';
-import MockOpenQClient from '../../services/ethers/MockOpenQClient';
+import MockOpenQClient from '../../../services/ethers/MockOpenQClient';
 describe('SetPayoutAdminPage', () => {
   const bounty = Constants.bounty1;
   beforeEach(() => {
@@ -39,7 +39,8 @@ describe('SetPayoutAdminPage', () => {
     // ASSERT
     expect(await screen.findByText(/Updating Payout.../)).toBeInTheDocument();
     expect(screen.getByText(/our request is being processed.../)).toBeInTheDocument();
-    expect(await screen.findByText(/updated/i)).toBeInTheDocument();
+    const updatedTexts = await screen.findAllByText(/updated/i);
+    expect(updatedTexts[0]).toBeInTheDocument();
     expect(await screen.findByText(/100.0 LINK/i)).toBeInTheDocument();
     expect(setPayout).toBeCalledWith(bounty.bountyId, '100', '0x5FbDB2315678afecb367f032d93F642f64180aa3');
   });
